Refactor to share kafka code
