--- server/common/src/main/scala/com/eigengo/lift/common/MessageDecoder.scala
@@ -1,10 +1,0 @@
-package com.eigengo.lift.common
-
-import com.eigengo.lift.common.MessagePayload.Payload
-
-import scalaz.\/
-
-trait MessageDecoder[A] {
-  def decode(data: Payload): String \/ A
-}
-
