--- server/common/src/main/scala/com/eigengo/lift/common/MessageEncoder.scala
@@ -1,9 +1,0 @@
-package com.eigengo.lift.common
-
-import com.eigengo.lift.common.MessagePayload.Payload
-
-import scalaz.\/
-
-trait MessageEncoder[A] {
-  def encode(value: A): String \/ Payload
-}
