+++ server/kafka-util/src/main/scala/com/eigengo/lift/kafkaUtil/MessageDecoder.scala
@@ -1,0 +1,10 @@
+package com.eigengo.lift.kafkaUtil
+
+import com.eigengo.lift.kafkaUtil.MessagePayload.Payload
+
+import scalaz.{\/, \/-, -\/}
+
+trait MessageDecoder[A] {
+  def decode(data: Payload): String \/ A
+}
+
