+++ server/kafka-util/src/main/scala/com/eigengo/lift/kafkaUtil/MessageEncoder.scala
@@ -1,0 +1,9 @@
+package com.eigengo.lift.kafkaUtil
+
+import com.eigengo.lift.kafkaUtil.MessagePayload.Payload
+
+import scalaz.{\/, \/-, -\/}
+
+trait MessageEncoder[A] {
+  def encode(value: A): String \/ Payload
+}
